+++ fluxengine_src/fe_setup_tools.py
@@ -12,7 +12,7 @@ This includes:
 @author: tomholding
 """
 
-from os import path, listdir;
+from os import path, listdir, getcwd, makedirs;
 import inspect;
 import time;
 import socket; #for gethostname
@@ -458,6 +458,113 @@ def fe_obj_from_run_parameters(runParameters, metadata, processLayersOff=True, v
         fe = None;
     
     return fe;
+
+#Takes a config file, a list of years and months, and runs the flux engine for each month/year combination.
+def run_fluxengine(configFilePath, yearsToRun, monthsToRun, verbose=False, processLayersOff=True,
+                   takahashiDriver=False, pco2DirOverride=None, outputDirOverride=None):
+    function = inspect.stack()[0][1]+", "+inspect.stack()[0][3];
+    hostname = socket.gethostname();
+    workingDirectory = getcwd();
+    if verbose:
+        print "Hostname identified as: ", hostname;
+        print "Working directory is: ", workingDirectory;
+    
+    global fe; #For debugging purposes this makes it easy to query the FluxEngine object after execution has finished.
+    
+    
+    #Parse config file
+    configPath = path.join(workingDirectory, configFilePath);
+    configVariables = read_config_file(configPath, verbose=verbose);
+    
+    #Parse settings file for default metadata about the config variables
+    settingsPath = path.join(workingDirectory, configVariables["src_home"], "settings.xml");
+    metadata = read_config_metadata(settingsPath, verbose=verbose);
+    
+    
+    #Substitute commandline override arguments (-pco2_dir_override, -output_dir_override)
+    if (pco2DirOverride != None):
+        print "Using optional override for pCO2w data directory. '%s' will be set to '%s' (ie overriding both directory of pco2 data and selection in the configuration file)." % (configVariables["pco2"], pco2DirOverride);
+        configVariables["pco2_path"] = path.abspath(pco2DirOverride);
+    if (outputDirOverride != None):
+        configVariables["output_dir"] = path.abspath(outputDirOverride);
+        print "Using optional override for output directory, output_dir will be set to % (overriding the configuration file value)." % (outputDirOverride);
+    
+    #Printing some feedback...
+    if processLayersOff == True and verbose:
+        print "Switching off generation of processing indicator layers. This reduces the processing time by appx. 50% (switch: process_layers_off).";
+        #configVariables["process_indicator_layers"] = None;
+    if takahashiDriver == True and verbose:
+        print "This is a takahashi validation run. Ensure config is the configs/takahashi09_validation.conf file supplied with FluxEngine.";
+    
+    #Using the metadata, process the config variables appropriately
+    #Checks types are valid, converts strings to the data types required by fluxengine
+    #Checks data layers contain at least a prod and path, etc.
+    verify_config_variables(configVariables, metadata, verbose=verbose);
+    
+    ############
+    ##TODO:
+    #Alert user to any unused variables (can we infer all required variables yet?).    
+    #Error if required variable hasn't been specified.
+    ############
+    
+    
+    #Begin main execution logic
+    processTimeStr = time.strftime("%d/%m/%Y %H:%M:%S");
+    print "Executing on '%s' at %s" % (hostname, processTimeStr);
+    
+    for year in yearsToRun:
+        for monthNum in monthsToRun:
+            #Run parameters can vary depending on the current month and year (e.g. paths and filenames,
+            #So these must be generated on a per-month/year basis.
+            try:
+                runParameters = create_run_parameters(configVariables, metadata, year, monthNum, processTimeStr, configFilePath, processLayersOff);
+                
+                #TODO: temporary stop-gap. Takahashi driver switch will be removed from future releases and moved to the configuration file.
+                if takahashiDriver == True:
+                    runParameters["TAKAHASHI_DRIVER"] = True;
+                else:
+                    runParameters["TAKAHASHI_DRIVER"] = False;
+            except ValueError as e:
+                print e.args;
+                return;
+            except OSError as e:
+                print e.args;
+                return;
+            
+            #Create output file path
+            try:
+                if path.exists(runParameters["output_dir"]) == False:
+                    makedirs(runParameters["output_dir"]);
+            except OSError as e:
+                print "Couldn't create output directory '%s'. Do you have write access?" % runParameters["output_dir"];
+                print type(e), e.args;
+            
+            #Create fluxengine object setup according to runParameters
+            fe = fe_obj_from_run_parameters(runParameters, metadata, processLayersOff, verbose=False);
+            
+            #Run fluxengine            
+            if fe != None:
+                #try:
+                    returnCode = fe.run();
+                #except Exception as e:
+                #    print "\n\n%s: Exception caught while running FluxEngine:" % function;
+                #    print type(e), e.args;
+                #    return 1;
+            
+            #Check for successful run, if one fails don't run the rest.
+            if returnCode != 0:
+                print ("%s: There was an error running flux engine:\n\n"%function), e.args[0];
+                print "Exiting...";
+                return returnCode;
+            else:
+                print "Flux engine exited with exit code:", returnCode;
+                print calendar.month_abbr[monthNum+1], year, "completed successfully.\n";
+            
+    #runStatus = {};
+    #runStatus["return_code"] = returnCode;
+    #runStatus["output_dir"] = runParameters["output_dir"];
+    #runStatus["config_used"] = configFilePath;
+    return 0;
 
 
 ##returns a dictionary the initialisation requirements for each rate parameterisation object (classes derived from KCalculationBase).
